feat(header): create dropdown nav component
